+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- New class `Youthweb\Api\Configuration` was added.
 - New interface `Youthweb\Api\Authentication\Authenticator` was added.
 - New class `Youthweb\Api\Authentication\NativeAuthenticator` was added.
 - The class `Youthweb\Api\Exception\ErrorResponseException` was added.
@@ -18,7 +19,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- **BREAKING:** The method `Youthweb\Api\Client::__construct()` was set to private, use `Youthweb\Api\Client::fromConfig()` instead.
 - **BREAKING:** The class `Youthweb\Api\Exception\UnauthorizedException` was set to `final`.
+- Replace `Cache\Adapter\Void\VoidCachePool` with new `Youthweb\Api\Cache\NullCacheItemPool` class.
 - Add parameter types and return types in nearly all classes
 - Change code style to PER coding style
 - Declare strict_types=1 in all PHP files
@@ -27,12 +30,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Removed
 
+- **BREAKING:** The method `Youthweb\Api\ClientInterface::__construct()` was removed.
 - **BREAKING:** The interface `Youthweb\Api\AuthenticatorInterface` was removed, use `Youthweb\Api\Authentication\Authenticator` instead.
 - **BREAKING:** The class `Youthweb\Api\YouthwebAuthenticator` was removed, use `Youthweb\Api\Authentication\NativeAuthenticator` instead.
 - **BREAKING:** The interface `Youthweb\Api\RequestFactoryInterface` was removed, use `Psr\Http\Message\RequestFactoryInterface` instead.
 - **BREAKING:** The class `Youthweb\Api\RequestFactory` was removed, use implementation of `Psr\Http\Message\RequestFactoryInterface` instead.
 - **BREAKING:** The interface `Youthweb\Api\HttpClientInterface` was removed, use `Psr\Http\Client\ClientInterface` instead.
 - **BREAKING:** The class `Youthweb\Api\HttpClient` was removed, use implementation of `Psr\Http\Client\ClientInterface` instead.
+- **BREAKING:** The class `Youthweb\Api\JsonObject` was removed.
 - Drop support for PHP 7.4
 
 ## [0.10.0](https://github.com/youthweb/php-youthweb-api/compare/0.9.0...0.10.0) - 2021-03-05